Move real time config manager to service directory
