+++ build-system/tasks/presubmit-checks.js
@@ -764,7 +764,6 @@ const forbiddenTerms = {
       'extensions/amp-a4a/0.1/test/test-a4a-var-source.js',
       'extensions/amp-a4a/0.1/test/test-amp-a4a.js',
       'extensions/amp-a4a/0.1/test/test-amp-ad-utils.js',
-      'extensions/amp-a4a/0.1/test/test-callout-vendors.js',
       'extensions/amp-a4a/0.1/test/test-refresh.js',
       'extensions/amp-access/0.1/test/test-access-expr.js',
       'extensions/amp-access/0.1/test/test-amp-login-done-dialog.js',
@@ -831,6 +830,7 @@ const forbiddenTerms = {
       'test/unit/test-amp-inabox.js',
       'test/unit/test-animation.js',
       'test/unit/test-batched-json.js',
+      'test/unit/test-callout-vendors.js',
       'test/unit/test-chunk.js',
       'test/unit/test-cid.js',
       'test/unit/test-css.js',
